+++ instrumentation/action_pack/test/opentelemetry/instrumentation/action_pack/handlers/action_controller_test.rb
@@ -16,7 +16,7 @@ describe OpenTelemetry::Instrumentation::ActionPack::Handlers::ActionController 
   let(:exporter) { EXPORTER }
   let(:spans) { exporter.finished_spans }
   let(:span) { exporter.finished_spans.last }
-  let(:rails_app) { DEFAULT_RAILS_APP }
+  let(:rails_app) { AppConfig.initialize_app }
   let(:config) { {} }
 
   # Clear captured spans
@@ -36,7 +36,6 @@ describe OpenTelemetry::Instrumentation::ActionPack::Handlers::ActionController 
 
     _(last_response.body).must_equal 'actually ok'
     _(last_response.ok?).must_equal true
-    _(span.name).must_equal 'ExampleController#ok'
     _(span.kind).must_equal :server
     _(span.status.ok?).must_equal true
 
@@ -65,7 +64,7 @@ describe OpenTelemetry::Instrumentation::ActionPack::Handlers::ActionController 
 
     _(last_response.body).must_equal 'created new item'
     _(last_response.ok?).must_equal true
-    _(span.name).must_equal 'ExampleController#new_item'
+    _(span.name).must_match(/^GET/)
     _(span.kind).must_equal :server
     _(span.status.ok?).must_equal true
 
@@ -82,24 +81,25 @@ describe OpenTelemetry::Instrumentation::ActionPack::Handlers::ActionController 
     _(span.attributes['code.function']).must_equal 'new_item'
   end
 
-  it 'sets the span name when the controller raises an exception' do
-    get 'internal_server_error'
+  describe 'when encountering server side errors' do
+    it 'sets semconv attributes' do
+      get 'internal_server_error'
 
-    _(span.name).must_equal 'ExampleController#internal_server_error'
-    _(span.kind).must_equal :server
-    _(span.status.ok?).must_equal false
+      _(span.kind).must_equal :server
+      _(span.status.ok?).must_equal false
 
-    _(span.instrumentation_library.name).must_equal 'OpenTelemetry::Instrumentation::Rack'
-    _(span.instrumentation_library.version).must_equal OpenTelemetry::Instrumentation::Rack::VERSION
+      _(span.instrumentation_library.name).must_equal 'OpenTelemetry::Instrumentation::Rack'
+      _(span.instrumentation_library.version).must_equal OpenTelemetry::Instrumentation::Rack::VERSION
 
-    _(span.attributes['http.method']).must_equal 'GET'
-    _(span.attributes['http.host']).must_equal 'example.org'
-    _(span.attributes['http.scheme']).must_equal 'http'
-    _(span.attributes['http.target']).must_equal '/internal_server_error'
-    _(span.attributes['http.status_code']).must_equal 500
-    _(span.attributes['http.user_agent']).must_be_nil
-    _(span.attributes['code.namespace']).must_equal 'ExampleController'
-    _(span.attributes['code.function']).must_equal 'internal_server_error'
+      _(span.attributes['http.method']).must_equal 'GET'
+      _(span.attributes['http.host']).must_equal 'example.org'
+      _(span.attributes['http.scheme']).must_equal 'http'
+      _(span.attributes['http.target']).must_equal '/internal_server_error'
+      _(span.attributes['http.status_code']).must_equal 500
+      _(span.attributes['http.user_agent']).must_be_nil
+      _(span.attributes['code.namespace']).must_equal 'ExampleController'
+      _(span.attributes['code.function']).must_equal 'internal_server_error'
+    end
   end
 
   it 'does not set the span name when an exception is raised in middleware' do
@@ -139,13 +139,79 @@ describe OpenTelemetry::Instrumentation::ActionPack::Handlers::ActionController 
     _(span.attributes['code.function']).must_be_nil
   end
 
+  describe 'span naming' do
+    describe 'when using the default span_naming configuration' do
+      describe 'successful requests' do
+        describe 'Rails Version < 7.1' do
+          it 'uses the http method controller and action name' do
+            skip "Rails #{Rails.gem_version} uses ActionDispatch::Request#route_uri_pattern" if Rails.gem_version >= Gem::Version.new('7.1')
+            get '/ok'
+
+            _(span.name).must_equal 'GET /example/ok'
+          end
+
+          it 'excludes route params' do
+            skip "Rails #{Rails.gem_version} uses ActionDispatch::Request#route_uri_pattern" if Rails.gem_version >= Gem::Version.new('7.1')
+            get '/items/1234'
+
+            _(span.name).must_equal 'GET /example/item'
+          end
+        end
+
+        describe 'Rails Version >= 7.1' do
+          it 'uses the Rails route' do
+            skip "Rails #{Rails.gem_version} does not define ActionDispatch::Request#route_uri_pattern" if Rails.gem_version < Gem::Version.new('7.1')
+            get '/ok'
+
+            _(span.name).must_equal 'GET /ok'
+          end
+
+          it 'includes route params' do
+            skip "Rails #{Rails.gem_version} does not define ActionDispatch::Request#route_uri_pattern" if Rails.gem_version < Gem::Version.new('7.1')
+            get '/items/1234'
+
+            _(span.name).must_equal 'GET /items/:id'
+          end
+        end
+      end
+
+      describe 'server errors' do
+        it 'uses the http method controller and action name for server side errors' do
+          skip "Rails #{Rails.gem_version} uses ActionDispatch::Request#route_uri_pattern" if Rails.gem_version >= Gem::Version.new('7.1')
+
+          get 'internal_server_error'
+
+          _(span.name).must_equal 'GET /example/internal_server_error'
+        end
+
+        it 'uses the Rails route for server side errors' do
+          skip "Rails #{Rails.gem_version} uses ActionDispatch::Request#route_uri_pattern" if Rails.gem_version < Gem::Version.new('7.1')
+
+          get 'internal_server_error'
+
+          _(span.name).must_equal 'GET /internal_server_error'
+        end
+      end
+    end
+
+    describe 'when using the class span_naming' do
+      let(:config) { { span_naming: :class } }
+
+      it 'uses the http method and controller name' do
+        get '/ok'
+
+        _(span.name).must_equal 'ExampleController#ok'
+      end
+    end
+  end
+
   describe 'when the application has exceptions_app configured' do
     let(:rails_app) { AppConfig.initialize_app(use_exceptions_app: true) }
 
     it 'does not overwrite the span name from the controller that raised' do
       get 'internal_server_error'
 
-      _(span.name).must_equal 'ExampleController#internal_server_error'
+      _(span.name).must_match(/^GET/)
       _(span.kind).must_equal :server
       _(span.status.ok?).must_equal false