+++ instrumentation/action_pack/test/test_helpers/app_config.rb
@@ -4,8 +4,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-class Application < Rails::Application; end
 require 'action_controller/railtie'
+class Application < Rails::Application; end
 require_relative 'middlewares'
 require_relative 'controllers'
 require_relative 'routes'
@@ -30,7 +30,7 @@ module AppConfig
     case Rails.version
     when /^6\.1/
       apply_rails_6_1_configs(new_app)
-    when /^7\./
+    when /^7|8\./
       apply_rails_7_configs(new_app)
     end