+++ instrumentation/action_pack/test/test_helpers/controllers/example_controller.rb
@@ -20,6 +20,6 @@ class ExampleController < ActionController::Base
   end
 
   def internal_server_error
-    raise :internal_server_error
+    raise 'internal_server_error'
   end
 end