+++ helpers/mysql/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-helpers-mysql
 
+### v0.1.2 / 2024-11-26
+
+* (No significant changes)
+
 ### v0.1.1 / 2024-06-18
 
 * FIXED: Relax otel common gem constraints