+++ instrumentation/active_record/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-instrumentation-active_record
 
+### v0.8.1 / 2024-11-21
+
+* FIXED: Pass block argument in ActiveRecord `find_by_sql` patch.
+
 ### v0.8.0 / 2024-10-22
 
 * BREAKING CHANGE: Rename Active Record find_by_sql spans to query